common: prplmesh_utils.sh: add non-intel option
Add an option to start the agent in non-intel mode. It simply passes the
-n option when starting the agent.

For the time being, this only applies to the agent. Controller can be
added later.

While we're at it, fix the getopt call: it was missing a colon after
iface-ctrl and iface-data to indicate that these options take an
argument.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>