agent: slave: Don't send VS TLV in WSC M1 if non-intel
We use the prplMesh agent to test the EasyMesh flows. However, prplMesh
agent still has the Intel Vendor-Specific messages and TLVs, so it
doesn't really act like an EasyMesh agent. The 'non-intel' option was
added to make it act like an EasyMesh agent.

For now, the only agent flow that is really implemented is the
autoconfig, where we add the agent information into the database based
on the M1 and radio basic info TLVs. Therefore, only the slave join is
excluded in the non-intel agent for the time being.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>