common: prplmesh_utils: fix check of slave operational
The `radio_agent_operational` function checks that the slave is
operational by looking for the string 'CSA' in the slave log file.
However, this is wrong for two reasons:

- CSA will also be printed in a few error paths.
- With a non-Intel agent, CSA will not be performed so the message never
  appears.

Therefore, look for the string 'goto STATE_OPERATIONAL' instead. As far
as the agent is concerned, this indicates that it is operational and
listens to controller commands.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>